int64

+++ others/measure_time_elapsed.cpp
@@ -1,7 +1,7 @@
 // To calculate the time taken by a code to execute
 #include <sys/time.h>
 
-long long getTimeInMicroseconds() {
+int64 getTimeInMicroseconds() {
     struct timeval start;
     gettimeofday(&start, NULL);
     return start.tv_sec * 1000000 + start.tv_usec;
@@ -11,7 +11,7 @@ long long getTimeInMicroseconds() {
 
 int main() {
     // write code
-    long long starttime = getTimeInMicroseconds();
+    int64 starttime = getTimeInMicroseconds();
     // sample(args) function run
     // Any other functions (if present) run
     std:cout << getTimeInMicroseconds() - starttime;
